+++ documentation/Changelog.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * `setCoupons` and `setShops` on project not index
-* implement `Equatable` and `Hashable` on project
+* implement `Equatable` and `Hashable` to Project
+* add `Hashable` to Coupon
 
 ### Added